Adds logo (#133)
This PR adds Vijil's logo to the documentation (on the top left corner).

I'm opening the PR just for the sake of using the staging deployments
because the following changes must be addressed:
1. The logo is too small. It must be increased a bit in size.
2. Mintlify allows users to switch between dark and light mode. For
these reasons, there must be two logos: one for the dark and the other
for the light mode

See the deployment on staging for understanding these reasons.